refactor: #180 core와 api 간 의존성 사용, 에러 가능성 있는 코드 보완
