+++ api/src/main/kotlin/com/inner/circle/api/security/MerchantApiKeyAuthenticationFilter.kt
@@ -1,18 +1,18 @@
 package com.inner.circle.api.security
 
-import com.inner.circle.core.security.CustomSecurityProvider
+import com.inner.circle.core.security.MerchantApiKeyProvider
 import com.inner.circle.exception.AppException
 import com.inner.circle.exception.HttpStatus
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import java.util.Base64
+import java.util.*
 import org.springframework.http.HttpHeaders
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
 
 class MerchantApiKeyAuthenticationFilter(
-    private val provider: CustomSecurityProvider
+    private val provider: MerchantApiKeyProvider
 ) : OncePerRequestFilter() {
     override fun doFilterInternal(
         request: HttpServletRequest,
@@ -29,15 +29,27 @@ class MerchantApiKeyAuthenticationFilter(
         filterChain.doFilter(request, response)
     }
 
-    private fun resolveApiKey(authHeader: String?): String {
+    private fun resolveApiKey(authHeader: String): String {
         authenticateWithBasicAuth(authHeader = authHeader)
 
-        val authorizationInfo = authHeader?.split(" ")
-        val encodedApiKey = authorizationInfo?.get(1)
-        val decodedApiKey = String(Base64.getDecoder().decode(encodedApiKey))
-        val apiKey = decodedApiKey.removeSuffix(":")
+        val authorizationInfo = authHeader.split(" ")
+        if (authorizationInfo.size != 2 || authorizationInfo[0] != "Basic") {
+            throw AppException(HttpStatus.UNAUTHORIZED, "Unauthorized: Basic Auth required")
+        }
 
-        return apiKey
+        val encodedApiKey = authorizationInfo[1]
+        val decodedApiKey =
+            try {
+                String(Base64.getDecoder().decode(encodedApiKey))
+            } catch (e: IllegalArgumentException) {
+                throw AppException(HttpStatus.UNAUTHORIZED, "Unauthorized: Basic Auth required")
+            }
+
+        if (!decodedApiKey.contains(":")) {
+            throw AppException(HttpStatus.UNAUTHORIZED, "Unauthorized: Basic Auth required")
+        }
+
+        return decodedApiKey.substringBefore(":")
     }
 
     private fun authenticateWithBasicAuth(authHeader: String?) {
