--- core/src/main/kotlin/com/inner/circle/core/security/CustomSecurityProvider.kt
@@ -1,7 +1,0 @@
-package com.inner.circle.core.security
-
-import org.springframework.security.core.Authentication
-
-interface CustomSecurityProvider {
-    fun getAuthentication(secret: String): Authentication
-}
