+++ core/src/main/kotlin/com/inner/circle/core/security/MerchantApiKeyProvider.kt
@@ -1,24 +1,7 @@
 package com.inner.circle.core.security
 
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.Authentication
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.stereotype.Component
 
-@Component
-class MerchantApiKeyProvider(
-    private val merchantDetailsService: MerchantDetailService
-) : CustomSecurityProvider {
-    override fun getAuthentication(token: String): Authentication {
-        val userDetails = merchantDetailsService.loadUserByUsername(token)
-        val authentication =
-            UsernamePasswordAuthenticationToken(
-                userDetails,
-                "",
-                listOf(
-                    SimpleGrantedAuthority("ROLE_MERCHANT")
-                )
-            )
-        return authentication
-    }
+interface MerchantApiKeyProvider {
+    fun getAuthentication(secret: String): Authentication
 }
