+++ core/src/main/kotlin/com/inner/circle/core/security/MerchantApiKeyProviderImpl.kt
@@ -1,0 +1,24 @@
+package com.inner.circle.core.security
+
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.authority.SimpleGrantedAuthority
+import org.springframework.stereotype.Component
+
+@Component
+class MerchantApiKeyProviderImpl(
+    private val merchantDetailsService: MerchantDetailService
+) : MerchantApiKeyProvider {
+    override fun getAuthentication(secret: String): Authentication {
+        val userDetails = merchantDetailsService.loadUserByUsername(secret)
+        val authentication =
+            UsernamePasswordAuthenticationToken(
+                userDetails,
+                "",
+                listOf(
+                    SimpleGrantedAuthority("ROLE_MERCHANT")
+                )
+            )
+        return authentication
+    }
+}
